check in

+++ softmax.c
@@ -13,9 +13,9 @@ static inline uint64_t rdtsc() {
     return a | ((uint64_t)d << 32);
 }
 
-void softmax(double *x, int length) {
+void softmax(float *x, int length) {
     
-    double max_val = x[0];
+    float max_val = x[0];
     
     // Find the maximum value in the array
     for (int i = 1; i < length; i++) {
@@ -30,7 +30,7 @@ void softmax(double *x, int length) {
 	}
 
     // Calculate the exponentials and sum
-    double sum = 0.0;
+    float sum = 0.0;
     for (int i = 0; i < length; i++) {
         sum += x[i];
     }
@@ -48,11 +48,11 @@ int main() {
     uint32_t clock, start, end;
     
     // Allocate memory for the array
-    double *x = (double *)malloc(length * sizeof(double));
+    float *x = (float *)malloc(length * sizeof(float));
 
     // Populate the array with random values
     for (int i = 0; i < length; i++) {
-        x[i] = ((double)rand() / RAND_MAX);
+        x[i] = ((float)rand() / RAND_MAX);
     }
 
     // Start count
